services/chat-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
